Sync world border animation & movement to Replay speed
Fixes an issue where the world border would move at normal speed when speeding up or slowing down the replay, and also at an incorrect speed when rendering the replay, causing it to get out of sync with the rest of the world.

I have only tested this fix for 1.17.1. My guess is that this will probably not work for versions below 1.14; I might fix that later.